src/requirements.txt.txt
tensorflow
sklearn
matplotlib
pyaml